extract out functions
